Bug fix: Indices in fixing and payment schedules were floats.

+++ models/hull_white/misc.py
@@ -396,8 +396,8 @@ def swap_schedule(fixing_start: int,
     # Number of fixings from fixing_start to fixing_end.
     n_fixings = (fixing_end - fixing_start) * fixing_frequency
     # Fixing and payment schedules.
-    fixing_schedule = np.zeros(n_fixings)
-    payment_schedule = np.zeros(n_fixings)
+    fixing_schedule = np.zeros(n_fixings, dtype=int)
+    payment_schedule = np.zeros(n_fixings, dtype=int)
     # Index of first fixing.
     start_idx = fixing_start * fixing_frequency * events_per_fixing
     for n in range(n_fixings):
